# multipedidos_api/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API Multipedidos",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cars": {
            "get": {
                "tags": [
                    "Carros"
                ],
                "summary": "Lista carros",
                "description": "Retorna uma lista com todos os carros cadastrados no banco de dados",
                "operationId": "158b081c3b62cbad1d47a7baf9ed52ac",
                "responses": {
                    "200": {
                        "description": "Lista carregada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Carros"
                ],
                "summary": "Cria um carro",
                "description": "Cria um novo carro de acordo com os atributos passados pelo corpo da requisição",
                "operationId": "581b89650bc67a7ae22937830482fcb7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand",
                                    "year",
                                    "description"
                                ],
                                "properties": {
                                    "model": {
                                        "type": "string",
                                        "example": "Gol"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Volkswagen"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2021
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Branco"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carro criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "get": {
                "tags": [
                    "Carros"
                ],
                "summary": "Exibe um carro",
                "description": "Exibe as informações do carro passado por parâmetro",
                "operationId": "d7bb47a818dc5b242dc4999f159d1037",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do carro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carro carregado com sucesso"
                    },
                    "404": {
                        "description": "Carro não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Carros"
                ],
                "summary": "Atualiza um carro",
                "description": "Atualiza as informações do carro passado por parâmetro",
                "operationId": "1feba9c073fea407739f3c5f99980c70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do carro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "brand",
                                    "year",
                                    "description"
                                ],
                                "properties": {
                                    "model": {
                                        "type": "string",
                                        "example": "Gol"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Volkswagen"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2021
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "Branco"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carro atualizado com sucesso"
                    },
                    "422": {
                        "description": "O valor informado não consta em nosso banco de dados"
                    },
                    "400": {
                        "description": "Carro não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carros"
                ],
                "summary": "Remove um carro",
                "description": "Remove o carro passado por parâmetro",
                "operationId": "8e2cc084cc4b07e94c769d0f92e214fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do carro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Carro removido com sucesso"
                    },
                    "404": {
                        "description": "Carro não encontrado"
                    }
                }
            }
        },
...
    }
}